% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgainStart.R
\docType{data}
\name{FAGUS}
\alias{FAGUS}
\title{Retained Carbon in Beech Trees}
\format{A data frame with 41 observations on the following 3 variables: 
\itemize{
\item \code{plot} (plot number)
\item \code{carbon} (carbon retained by leaves measured in kg/ha)
\item \code{type} (a factor with levels \code{monospecific} and \code{multispecific})
}}
\source{
\emph{Gobierno de Navarra} and \emph{Gestion Ambiental Viveros y Repoblaciones de Navarra}, 2006.  The data were obtained within the European Project FORSEE.
}
\usage{
FAGUS
}
\description{
The carbon retained by leaves measured in kg/ha is recorded for forty-one different plots of mountainous regions of Navarre (Spain), depending on the forest classification: areas with 90\% or more beech trees (\emph{Fagus Sylvatica}) are labeled monospecific, while areas with many species of trees are labeled multispecific.
}
\examples{
ggplot(data = FAGUS, aes(x = type, y = carbon)) + geom_boxplot()
}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. \emph{Probability and Statistics with R}, Second Edition. Chapman & Hall / CRC.
}
\keyword{datasets}

