% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgainStart.R
\docType{package}
\name{PASWR2-package}
\alias{PASWR2-package}
\title{The PASWR2 Package}
\description{
The PASWR2 Package
}
\details{
Data sets and functions for \emph{Probability and Statistics with R}, Second Edition.
}
\keyword{package}

