% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgainStart.R
\docType{data}
\name{SALINITY}
\alias{SALINITY}
\title{Surface-Water Salinity}
\format{
A data frame with 48 observations on the following variable: 
\itemize{
\item \code{salinity} (surface-water salinity measurements)
}
}
\source{
Davis, J. 1986. \emph{Statistics and Data Analysis in Geology}. John Wiley, New York.
}
\usage{
SALINITY
}
\description{
Surface-water salinity measurements were taken in a bottom-sampling project in Whitewater Bay, Florida.
}
\examples{
# Boxplot
ggplot(data = SALINITY, aes(x = salinity)) + geom_density(fill = "yellow", alpha = 0.3)
}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. \emph{Probability and Statistics with R}, Second Edition. Chapman & Hall / CRC.
}
\keyword{datasets}
