% Generated by roxygen2 (4.0.2): do not edit by hand
\name{srs}
\alias{srs}
\title{Simple Random Sample}
\usage{
srs(popvalues, n)
}
\arguments{
\item{popvalues}{are values of the population.  \code{NA}s and \code{Inf}s are allowed but will be removed from the population.}

\item{n}{the sample size}
}
\value{
The function \code{srs()} returns a matrix containing the posible simple random samples of size \code{n} taken from a population of finite values \code{popvalues}.
}
\description{
Computes all possible samples from a given population using simple random sampling
}
\details{
If non-finite values are entered as part of the population, they are removed; and the returned simple random sample computed is based on the remaining finite values.
}
\examples{
srs(popvalues = c(5, 8, 3, NA, Inf), n = 2)
}
\author{
Alan T. Arnholt <arnholtat@appstate.edu>
}
\seealso{
\code{\link{combn}}
}
\keyword{programming}

