% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgainStart.R
\docType{data}
\name{STATTEMPS}
\alias{STATTEMPS}
\title{Student Temperatures}
\format{A data frame with 34 observations on the following 3 variables: 
\itemize{
\item \code{temperature} (temperature in Fahrenheit)
\item \code{gender} (a factor with levels \code{Female} and \code{Male})
\item \code{class} (a factor with levels \code{8 a.m.} and \code{9 a.m.})
}}
\usage{
STATTEMPS
}
\description{
In a study conducted at Appalachian State University, students used digital oral thermometers to record their temperatures each day they came to class.  A randomly selected day of student temperatures is provided in \code{STATTEMPS}.  Information is also provided with regard to subject gender and the hour of the day when the students' temperatures were measured.
}
\examples{
p <- ggplot(data = STATTEMPS, aes(x = gender, y =temperature, fill = class))
p + geom_violin()
}
\references{
Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. \emph{Probability and Statistics with R}, Second Edition. Chapman & Hall / CRC.
}
\keyword{datasets}

