% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TITANIC3}
\alias{TITANIC3}
\title{Titanic Survival Status}
\format{A data frame with 1309 observations on the following 14 variables:
\itemize{
\item \code{pclass} (a factor with levels \code{1st}, \code{2nd}, and \code{3rd})
\item \code{survived} (Survival wehre 0 = No; 1 = Yes)
\item \code{name} (Name)
\item \code{sex} (a factor with levels \code{female} and \code{male})
\item \code{age} (age in years)
\item \code{sibsp} (Number of Siblings/Spouses Aboard)
\item \code{parch} (Number of Parents/Children Aboard)
\item \code{ticket} (Ticket Number)
\item \code{fare} (Passenger Fare)
\item \code{cabin} (Cabin)
\item \code{embarked} (a factor with levels \code{Cherbourg}, \code{Queenstown}, and \code{Southampton})
\item \code{boat} (Lifeboat Number)
\item \code{body} (Body Identification Number)
\item \code{home.dest} (Home/Destination)
}}
\source{
\url{http://biostat.mc.vanderbilt.edu/twiki/pub/Main/DataSets/titanic.html}
}
\usage{
TITANIC3
}
\description{
The \code{TITANIC3} data frame describes the survival status of individual passengers on the Titanic.  The  \code{TITANIC3} data frame does not contain information for the crew, but it does contain actual and estimated ages for almost 80\% of the passengers.
}
\details{
Thomas Cason from the University of Virginia has greatly updated and improved the \code{titanic} data frame using the \emph{Encyclopedia Titanica} and created a new dataset called \code{TITANIC3}. This dataset reflects the state of data available as of August 2, 1999. Some duplicate passengers have been dropped; many errors have been corrected; many missing ages have been filled in; and new variables have been created.
}
\examples{
with(TITANIC3, table(pclass, sex))
}
\references{
\itemize{
\item Harrell, F. E. 2001.
 \emph{Regression Modeling Strategies with Applications to Linear Models, Logistic Regression, and Survival Analysis}.  Springer.
\item Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. \emph{Probability and Statistics with R}, Second Edition. Chapman & Hall / CRC.
}
}
\keyword{datasets}

